requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.1
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==64.2.1
django-cors-headers==4.4.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
simplejson==3.19.3
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
